Cleanup Fusion rewrites
* Move local_add_mul_fusion to `rewriting/elemwise` and remove unused/duplicated TestAddMulFusion tests
* Use EquilibriumGraphRewriter for local_add_mul_fusion
* Do not register optional rewrites if tensor__local_elemwise_fusion flag is disabled
